==> notebooks/predict_infection/1 - Generate Dataset.ipynb <==
import sys
sys.path.append("../../src/models")

import numpy as np
from actions import city_restrictions, costs

from simulate_pandemic import init_infection, spread_infection, lambda_leak_expose, update_population
from simulate_pandemic import main

from numpy.random import default_rng
import pandas as pd
from datetime import datetime
from tqdm import tqdm
from joblib import Parallel, delayed

## Generate simulation data

def make_individual_df(week, sim, action, data):
    df = pd.DataFrame(data, columns=['id', 'state'])
    df['simulation'] = sim
    df['week'] = week
    df['action'] = action
    
    return df

def make_value_counts_df(week, sim, action, data):
    df = pd.DataFrame(pd.Series(data[:, 1]).value_counts()).T
    df['simulation'] = sim
    df['week'] = week
    df['action'] = action
    df = df.rename(columns = {
                              -1 : 'removed',
                               0 : 'susceptible',
                               1 : 'exposed',
                               2 : 'infected',
                               3 : 'hospitalized'
                              }
                  )
    
    return df

def simulate_and_save(policy_number, policy, n_sims, step_size=7, folder='generated_sims/'):
    dfs = []
    for sim in range(1, n_sims+1):
        
        sim_name = f'{policy_number}_{sim}'
                
        data, pop_matrix = main(gpickle_path=gpickle_path,
                                p_r=p_r,
                                policy=policy,
                                disable_tqdm=True,
                                days=step_size*len(policy),
                                step_size=step_size)
        
        weeks = (step_size/7) * len(policy)
        
        weekly_data = [data[i*7] for i in range(int(weeks))]
        df = pd.concat([make_individual_df(week, sim_name, action, data)
                             for week, (data, action) in enumerate(zip(weekly_data, policy))])
        #df.to_parquet(folder + f'{sim_name}.parquet')
        dfs += [df]
    return dfs

rng = default_rng(None)

gpickle_path = "../../data/processed/SP_multiGraph_Job_Edu_Level.gpickle"

prhome = 0.06
p_r = {
    'home'    :  prhome,
    'neighbor':  .1*prhome,
    'work'    :  .1*prhome,
    'school'  :  .15*prhome,
}

costs_keys = list(costs.keys())
costs_values = np.array(list(costs.values()))


### Cost Weighted

p = (1 - costs_values) / (1 - costs_values).sum()
policy = [rng.choice(costs_keys, size=int(364/14),
                     replace=True, p=p)
          for i in range(16)]

assert len(set([tuple(p) for p in policy])) == len(policy)

n_sims=1
res = Parallel(n_jobs=16)(delayed(simulate_and_save)(i, pol, n_sims) 
                              for i, pol in tqdm(enumerate(policy),
                                                 total=len(policy)))

res_df = pd.concat([r for rr in res for r in rr])

res_df.shape

res_df.to_parquet("simulation_results_dataset.parquet")

## Add individuals data

individual_infos = pd.read_feather("../../data/processed/clusterized_df.feather")

individual_infos

dataset_info = pd.merge(
    res_df,
    individual_infos[['id', 'home', 'school', 'work', 'idade', 'home_id']].astype(float),
    on='id'
).copy()

## Creating Target (State two weeks from now)

dataset_info['target'] = dataset_info.groupby(['id', 'simulation'])['state'].shift(-1)

dataset_info.groupby(['week'])['state'].apply(lambda x: (x == 0).sum() / len(x)).mean()

### Discarding weeks 17 onwards, due to 30% removed

dataset_info = dataset_info[dataset_info['week'] < 17]

dataset_info.target.isna().sum()

dataset_info.query("id == 1936 and week == 12 and simulation == '0_1'")

dataset_info['target'] = dataset_info['target'].astype(int)

def create_columns_map(relation_type):
    return {
        'week': 'actual_week',
        -1: f'{relation_type}_removed',
        0: f'{relation_type}_susceptible',
        1: f'{relation_type}_exposed',
        2: f'{relation_type}_infected',
        3: f'{relation_type}_hospitalized'
    }

resulting_ratios = {}

for relation_type in ['home', 'school', 'work', 'home_id']:
    print(relation_type)
    value_counts_week = (
        dataset_info
            .groupby(['week', 'simulation', relation_type])['target']
            .value_counts(normalize=True)
            .reset_index(name='ratio')
    )
    
    print("value counts done")
    pivot = (
        pd.pivot_table(data=value_counts_week, index=['week', 'simulation', relation_type],
                       columns='target', values='ratio', fill_value=0)
          .reset_index()
          .rename(columns=create_columns_map(relation_type))
    )
    
    print("Pivot Done")
   
    pivot['week'] = pivot['actual_week'] + 1
    
    resulting_ratios[relation_type] = pivot

model_dataset = dataset_info[(dataset_info['target'].notna()) & (dataset_info['state'] == 0)].copy()

model_dataset.shape[0] / dataset_info.shape[0]

model_dataset.query("id == 1936 and week == 12 and simulation == '0_1'")

before_shape = model_dataset.shape[0]

model_dataset = pd.merge(
    model_dataset,
    resulting_ratios['home'].drop("actual_week", axis=1),
    on=['week', 'simulation', 'home'],
    how='left'
).merge(
    resulting_ratios['work'].drop("actual_week", axis=1),
    on=['week', 'simulation', 'work'],
    how='left'
).merge(
    resulting_ratios['school'].drop("actual_week", axis=1),
    on=['week', 'simulation', 'school'],
    how='left'
).merge(
    resulting_ratios['home_id'].drop("actual_week", axis=1),
    on=['week', 'simulation', 'home_id'],
    how='left'
)

assert model_dataset.shape[0] == before_shape

model_dataset['numerical_actions'] = model_dataset['action'].map({
    'Unrestricted': 0,
    'Social Distancing':1,
    'Light Quarantine': 2,
    'Hard Quarantine':3,
    'Lockdown':4
})

Should we consider only target state 1 or any target state different than 0? i.e., the person can be hospitalized, should also count as a one.

model_dataset['binary_target'] = (model_dataset['target'] != 0).astype(int)

model_dataset['train_test'] = list(map({True: 'train', False: 'test'}.get, (model_dataset['week'] < 11)))

model_dataset.groupby(['train_test', 'week'])['binary_target'].mean().mean()

model_dataset.reset_index(drop=True).to_parquet("model_dataset.parquet")

model_dataset.shape
